Refactor Foundation section structure and add OrganizationalStructure component with styles
