+++ client/src/views/Foundation/FoundationAtv/FoundationATV.jsx
@@ -1,20 +1,20 @@
 import { useState, useRef } from "react";
 import "boxicons";
-import FotoVideo from "../../../public/img_Foundation/foto-video.jpg";
-import Foundation1 from "../../../public/img_Foundation/foundation-1.jpg";
-import Foundation2 from "../../../public/img_Foundation/foundation-2.jpg";
-import BannerFoundation from "../../assets/images/views/foundation/banner-foundation.webp";
-import TeoriaCambio from "../../assets/images/views/foundation/entrega-pc.webp";
-import Valores from "../../assets/images/views/foundation/valores.webp";
-import Mision from "../../assets/images/views/foundation/pase-a-la-u.webp";
-import Publico from "../../assets/images/views/foundation/estudiantes-colegio.webp";
-import Vision from "../../assets/images/views/foundation/estudiantes-universitarios.webp";
-import Proposito from "../../assets/images/views/foundation/foundation-estudiantes.webp";
-import { Sun, SunBlue } from "../../assets/images/svg/Svg";
+import FotoVideo from "../../../../public/img_Foundation/foto-video.jpg";
+import Foundation1 from "../../../../public/img_Foundation/foundation-1.jpg";
+import Foundation2 from "../../../../public/img_Foundation/foundation-2.jpg";
+import BannerFoundation from "../../../assets/images/views/foundation/banner-foundation.webp";
+import TeoriaCambio from "../../../assets/images/views/foundation/entrega-pc.webp";
+import Valores from "../../../assets/images/views/foundation/valores.webp";
+import Mision from "../../../assets/images/views/foundation/pase-a-la-u.webp";
+import Publico from "../../../assets/images/views/foundation/estudiantes-colegio.webp";
+import Vision from "../../../assets/images/views/foundation/estudiantes-universitarios.webp";
+import Proposito from "../../../assets/images/views/foundation/foundation-estudiantes.webp";
+import { Sun, SunBlue } from "../../../assets/images/svg/Svg";
 
 import { useTranslation } from "react-i18next";
 import { motion, useInView } from "framer-motion";
-import BannerView from "../Banner-views/BannerView";
+import BannerView from "../../../components/Banner-views/BannerView";
 
 function FoundationATV() {
   const { t } = useTranslation();
